+++ README.md
@@ -18,7 +18,7 @@ Import a subpath:
 
 ```js
 // ESM / Typescript
-import { pascalCase  } from 'unkit/string'
+import { pascalCase } from 'unkit/string'
 
 // CommonJS
 const { pascalCase } = require('unkit/string')
@@ -38,6 +38,7 @@ Subpath | Packages
 [`unkit/esm`](#esm)   | [unjs/mlly](https://github.com/unjs/mlly)
 [`unkit/fetch`](#fetch)  | [unjs/ohmyfetch](https://github.com/unjs/ohmyfetch)
 [`unkit/http`](#http)  | [unjs/h3](https://github.com/unjs/h3), [unjs/is-https](https://github.com/unjs/is-https)
+[`unkit/nuxt`](#nuxt) | [Nuxt Module](https://nuxt.com/modules)
 [`unkit/object`](#object)  | [unjs/defu](https://github.com/unjs/defu), [unjs/destr](https://github.com/unjs/destr)
 [`unkit/promise`](#promise)  | [unjs/items-promise](https://github.com/unjs/items-promise)
 [`unkit/string`](#string)  | [unjs/scule](https://github.com/unjs/scule)
@@ -94,6 +95,17 @@ import { useBody, isHTTPS } from 'unkit/http'
 
 // CommonJS
 const { useBody, isHTTPS } = require('unkit/http')
+```
+
+### `/nuxt`
+
+> Enables auto-imports for utilities from unkit
+
+```ts
+export default defineNuxtConfig({
+  modules: ['unkit/nuxt'],
+  // unkit: {}
+})
 ```
 
 ### `/object`
