+++ build.config.ts
@@ -1,0 +1,5 @@
+import { defineBuildConfig } from 'unbuild'
+
+export default defineBuildConfig({
+  externals: ['@nuxt/kit', '@nuxt/schema']
+})
