--- playground/nuxt.config.ts
@@ -1,9 +1,0 @@
-// import UnkitModule from '../nuxt'
-
-export default defineNuxtConfig({
-  alias: { unkit: '../src' },
-  modules: ['../src/nuxt'],
-  unkit: {
-    prefix: 'useUnJS'
-  }
-})
