--- playground/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "unkit-playground",
-  "private": true
-}
