--- src/nuxt/imports.ts
@@ -1,8 +1,0 @@
-export * as env from '../env' // 'unkit/env' or '~/src/env'
-export * as esm from '../esm'
-export * as fetch from '../fetch'
-export * as http from '../http'
-export * as object from '../object'
-export * as promise from '../promise'
-export * as string from '../string'
-export * as url from '../url'
